requirements.txt
confusable-homoglyphs==3.2.0
Django==1.11.15
django-registration==2.4.1
pep8==1.7.1
pytz==2018.9
bootstrap4==0.0.5
dj-database-url==0.5.0
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.7.5
python-decouple==3.1
whitenoise==3.3.1
config==0.4.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-location-field==2.0.6
django-tinymce==2.7.0
Pillow==5.3.0
psycopg2-binary==2.7.5
six==1.11.0

